Successfully registered SDK! Also cleaned up Android Manifest
The connection activity has been successfully implemented and will likely be used in some form in the final version of the app. Likewise the camera screen could be integrated. They are currently under options but are anticipated to move. Furthermore, the manifest has been formatted properly and the camera screen now correctly shifts to a landscape view.